+++ system/modules/networking.nix
@@ -4,6 +4,9 @@
   networking.nftables.enable = false;
   networking.firewall.enable = false;
   networking.hosts = {
-    "127.0.0.1" = [ "notebook.local" ];
+    "127.0.0.1" = [ 
+      "notebook.local"
+      "immich.local" 
+    ];
   };
 }